refactor(glass-core): distribute session tests beside their code (Phase B)
Split the 2,365-line session test module: each of the 71 tests moves into a
`#[cfg(test)] mod tests` in the submodule it exercises (a11y 23, wait 21,
lifecycle 7, capture 5, baseline 4, clipboard 3, core 4, input 2, window 2).
Shared scaffolding (FakePlatform/FakeAccessibility + glass_with* builders)
moves to a `#[cfg(test)]` `session::test_support` module, exposed pub(crate).
The a11y-tree helper free fns move from mod.rs to a11y.rs alongside their tests.

Tests moved verbatim (no assertion changed). glass-core test count unchanged
at 214 (identical before/after), proving behavior preserved.

Verified: clippy -p glass-core --all-targets -D warnings, test --workspace, fmt --check all green.
