+++ crates/glass-core/src/session/window.rs
@@ -42,3 +42,65 @@ impl Glass {
         Ok(geometry)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::session::test_support::*;
+
+    #[test]
+    fn window_resize_updates_tracked_geometry() {
+        let mut g = glass_with(FakePlatform::new(10, 10));
+        g.start(&spec()).unwrap();
+        let geom = g
+            .window(&WindowOp::Resize {
+                width: 20,
+                height: 30,
+            })
+            .unwrap();
+        assert_eq!(geom.width, 20);
+        assert_eq!(geom.height, 30);
+        assert_eq!(g.geometry().unwrap().width, 20);
+    }
+
+    #[test]
+    fn select_window_switches_active_geometry() {
+        let a = WindowInfo {
+            id: WindowId(1),
+            title: Some("A".into()),
+            class: None,
+            geometry: WindowGeometry {
+                x: 0,
+                y: 0,
+                width: 320,
+                height: 240,
+            },
+            active: true,
+        };
+        let b = WindowInfo {
+            id: WindowId(2),
+            title: Some("B".into()),
+            class: None,
+            geometry: WindowGeometry {
+                x: 400,
+                y: 0,
+                width: 100,
+                height: 80,
+            },
+            active: false,
+        };
+        let mut glass = glass_with(FakePlatform::new(320, 240).with_windows(vec![a, b]));
+        glass.start(&spec()).unwrap();
+
+        let listed = glass.list_windows().unwrap();
+        assert_eq!(listed.len(), 2);
+
+        let geo = glass.select_window(WindowId(2)).unwrap();
+        assert_eq!((geo.width, geo.height), (100, 80));
+        assert_eq!(glass.geometry().unwrap().width, 100);
+
+        assert!(matches!(
+            glass.select_window(WindowId(999)),
+            Err(GlassError::WindowNotFound)
+        ));
+    }
+}
